@@ -11,39 +11,40 @@
     },
     "license": "ISC",
     "dependencies": {
-        "@angular/animations": "^4.4.6",
-        "@angular/common": "^4.4.6",
-        "@angular/compiler": "^4.4.6",
-        "@angular/core": "^4.4.6",
-        "@angular/forms": "^4.4.6",
-        "@angular/http": "^4.4.6",
-        "@angular/platform-browser": "^4.4.6",
-        "@angular/platform-browser-dynamic": "^4.4.6",
-        "@angular/router": "^4.4.6",
+        "@angular/animations": "^5.0.0",
+        "@angular/common": "^5.0.0",
+        "@angular/compiler": "^5.0.0",
+        "@angular/core": "^5.0.0",
+        "@angular/forms": "^5.0.0",
+        "@angular/http": "^5.0.0",
+        "@angular/platform-browser": "^5.0.0",
+        "@angular/platform-browser-dynamic": "^5.0.0",
+        "@angular/router": "^5.0.0",
         "core-js": "^2.4.1",
-        "rxjs": "^5.5.0",
-        "zone.js": "^0.8.18",
+        "rxjs": "^5.5.2",
+        "zone.js": "^0.8.14",
 
         "apollo-client":"^1.9.0",
         "apollo-angular":"^0.13.0",
         "graphql-tag":"^2.5.0",
 
-        "@angular/flex-layout": "^2.0.0-beta.9",
-        "@angular/material": "^2.0.0-beta.12",
-        "@angular/cdk": "^2.0.0-beta.12",
+        "@angular/flex-layout": "^2.0.0-beta.11",
+        "@angular/material": "^5.0.0-rc0",
+        "@angular/cdk": "^5.0.0-rc0",
         "hammerjs": "^2.0.8",
 
         "moment": "^2.0.0"
     },
     "devDependencies": {
-        "@angular/cli": "^1.4.9",
-        "@angular/compiler-cli": "^4.2.4",
-        "@angular/language-service": "^4.2.4",
+
+        "@angular/cli": "^1.5.0",
+        "@angular/compiler-cli": "^5.0.0",
+        "@angular/language-service": "^5.0.0",
         "@types/hammerjs": "^2.0.0",
         "@types/jasmine": "~2.5.53",
         "@types/jasminewd2": "~2.0.2",
-        "@types/node": "~6.0.90",
-        "codelyzer": "~3.1.1",
+        "@types/node": "~6.0.60",
+        "codelyzer": "~3.2.0",
         "jasmine-core": "~2.6.2",
         "jasmine-spec-reporter": "~4.1.0",
         "karma": "~1.7.0",
@@ -54,8 +55,8 @@
         "karma-jasmine-html-reporter": "^0.2.2",
         "protractor": "~5.1.2",
         "ts-node": "~3.2.0",
-        "tslint": "~5.3.2",
-        "typescript": "~2.3.0"
+        "tslint": "~5.7.0",
+        "typescript": "~2.4.2"
     },
     "repository": {}
 }
