@@ -1,8 +1,8 @@
 import {HashLocationStrategy, LocationStrategy} from '@angular/common';
+import {HttpClientModule} from '@angular/common/http';
 import {NgModule} from '@angular/core';
 import {FlexLayoutModule} from '@angular/flex-layout';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {HttpModule} from '@angular/http';
 import {DateAdapter, NativeDateAdapter} from '@angular/material';
 import {BrowserModule} from '@angular/platform-browser';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
@@ -13,12 +13,12 @@ import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {AuthModule} from './auth/auth.module';
 import {ApolloFactoryLoader, ApolloService, getApolloClient} from './core/apollo.service';
+import {AuthService} from './core/auth.service';
 import {CoreModule} from './core/core.module';
 import {DashboardComponent} from './dashboard/dashboard.component';
 import {LoginComponent} from './login/login.component';
 import {MainComponent} from './main/main.component';
 import './rxjs-operators';
-import {AuthService} from './core/auth.service';
 import {ConfirmationDialogComponent} from './shared/component/dialog/confirmation-dialog/confirmation-dialog.component';
 import {SharedModule} from './shared/shared.module';
 
@@ -28,7 +28,7 @@ import {SharedModule} from './shared/shared.module';
         BrowserAnimationsModule,
         FormsModule,
         ReactiveFormsModule,
-        HttpModule,
+        HttpClientModule,
 
         CoreModule,
         AppRoutingModule,
